+++ core/src/utils/test/playwright/page/event-spy.ts
@@ -1,3 +1,5 @@
+import type { JSHandle } from '@playwright/test';
+
 import type { E2EPage } from '../playwright-declarations';
 
 /**
@@ -16,12 +18,20 @@ export class EventSpy {
    */
   private cursor = 0;
   private queuedHandler: (() => void)[] = [];
-  public events: Event[] = [];
+  public events: CustomEvent[] = [];
 
   constructor(public eventName: string) {}
 
   get length() {
     return this.events.length;
+  }
+
+  get firstEvent() {
+    return this.events[0] || null;
+  }
+
+  get lastEvent() {
+    return this.events[this.events.length - 1] || null;
   }
 
   public next() {
@@ -48,7 +58,7 @@ export class EventSpy {
     }
   }
 
-  public push(ev: Event) {
+  public push(ev: CustomEvent) {
     this.events.push(ev);
     const next = this.queuedHandler.shift();
     if (next) {
@@ -82,17 +92,69 @@ export const initPageEvents = async (page: E2EPage) => {
  * page context to updates the _e2eEvents map
  * when an event is fired.
  */
-export const addE2EListener = async (page: E2EPage, eventName: string, callback: (ev: any) => void) => {
+export const addE2EListener = async (
+  page: E2EPage,
+  elmHandle: JSHandle,
+  eventName: string,
+  callback: (ev: any) => void
+) => {
   const id = page._e2eEventsIds++;
   page._e2eEvents.set(id, {
     eventName,
     callback,
   });
 
-  await page.evaluate(
-    ([eventName, id]) => {
-      window.addEventListener(eventName as string, (ev: Event) => {
-        (window as any).ionicOnEvent(id, ev);
+  await elmHandle.evaluate(
+    (elm, [eventName, id]) => {
+      (window as any).stencilSerializeEventTarget = (target: any) => {
+        // BROWSER CONTEXT
+        if (!target) {
+          return null;
+        }
+        if (target === window) {
+          return { serializedWindow: true };
+        }
+        if (target === document) {
+          return { serializedDocument: true };
+        }
+        if (target.nodeType != null) {
+          const serializedElement: any = {
+            serializedElement: true,
+            nodeName: target.nodeName,
+            nodeValue: target.nodeValue,
+            nodeType: target.nodeType,
+            tagName: target.tagName,
+            className: target.className,
+            id: target.id,
+          };
+          return serializedElement;
+        }
+        return null;
+      };
+
+      (window as any).serializeStencilEvent = (orgEv: CustomEvent) => {
+        const serializedEvent = {
+          bubbles: orgEv.bubbles,
+          cancelBubble: orgEv.cancelBubble,
+          cancelable: orgEv.cancelable,
+          composed: orgEv.composed,
+          currentTarget: (window as any).stencilSerializeEventTarget(orgEv.currentTarget),
+          defaultPrevented: orgEv.defaultPrevented,
+          detail: orgEv.detail,
+          eventPhase: orgEv.eventPhase,
+          isTrusted: orgEv.isTrusted,
+          returnValue: orgEv.returnValue,
+          srcElement: (window as any).stencilSerializeEventTarget(orgEv.srcElement),
+          target: (window as any).stencilSerializeEventTarget(orgEv.target),
+          timeStamp: orgEv.timeStamp,
+          type: orgEv.type,
+          isSerializedEvent: true,
+        };
+        return serializedEvent;
+      };
+
+      elm.addEventListener(eventName as string, (ev: Event) => {
+        (window as any).ionicOnEvent(id, (window as any).serializeStencilEvent(ev));
       });
     },
     [eventName, id]