+++ core/src/utils/test/playwright/page/utils/index.ts
@@ -4,3 +4,4 @@ export * from './get-snapshot-settings';
 export * from './set-ion-viewport';
 export * from './spy-on-event';
 export * from './set-content';
+export * from './locator';